% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_area.R
\name{calculate_area}
\alias{calculate_area}
\title{calculate_area}
\usage{
calculate_area(d, digits = NULL)
}
\arguments{
\item{d}{A numeric value for length (in kilometres) of the maximum distance
of a village/community from a sampling point.}

\item{digits}{A numeric value specifying the number of digits to round off
the resulting area size. Default is NULL}
}
\value{
A data.frame of 1 row and 2 columns. First column is for the area
    (in square kilometres) for a triangle. Second column is for the area (in
    square kilometres) for a hexagon.
}
\description{
Function to calculate the area size of the resulting hexagon in a hexagonal
grid and the area size of the resulting triangle in a triangular grid for a
specified value of \code{d} (in kilometres) used in the simple spatial
sampling method (S3M).
}
\examples{
# Calculate the area of the triangle and hexagon for a d of 10 kilometres
calculate_area(d = 10, digits = 2)

}
\author{
Ernest Guevarra based on equations by Mark Myatt
}
