% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_d.R
\name{calculate_d}
\alias{calculate_d}
\title{calculate_d}
\usage{
calculate_d(area, geom = c("tri", "hex"))
}
\arguments{
\item{area}{Area (square kilometres) of a hexagon in a hexagonal grid.
Determines the spatial resolution of the resulting spatial sample.}

\item{geom}{Spatial geometry to which \code{area} is specified for. Can be
one of two options: \code{"tri"} for a triangle; \code{hex} for a
hexagon.}
}
\value{
A numeric value for length (kilometres) of \code{d}
}
\description{
Function to calculate the distance \code{d} given an area size (in square
kilometres) of a hexagon in a hexagonal spatial grid determining spatial
resolution of spatial sampling
}
\examples{
# Calculate d for a triangle with an area size of 130 square kilometres
calculate_d(area = 130, geom = "tri")

# Calculate d for a hexagon with an area size of 260 square kilometres
calculate_d(area = 260, geom = "hex")

}
\author{
Ernest Guevarra based on equations by Mark Myatt
}
