% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_height.R
\name{calculate_height}
\alias{calculate_height}
\title{Calculate height (distance from north to south) of S3M grid given a value of
\code{d}}
\usage{
calculate_height(d)
}
\arguments{
\item{d}{A numeric value for length (in kilometres) of the maximum distance
of a village/community from a sampling point. This can be the output of
the function \code{calculate_d()}}
}
\value{
A numeric value in kilometres of height (distance from north to
    south) of the S3M grid
}
\description{
Calculate height (distance from north to south) of S3M grid given a value of
\code{d}
}
\examples{

# Calculate height given a d of 6 kms
calculate_height(d = 6)

}
\author{
Ernest Guevarra based on equations by Mark Myatt
}
