% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_length.R
\name{calculate_length}
\alias{calculate_length}
\title{calculate_length}
\source{
Based on equations by Mark Myatt
}
\usage{
calculate_length(d)
}
\arguments{
\item{d}{A numeric value for length (in kilometres) of the maximum distance
of a village/community from a sampling point. This can be the output of
the function \code{calculate_d()}}
}
\value{
A numeric value in kilometres of length (distance from east to west)
    of the S3M grid
}
\description{
Calculate length (distance from east to west) of S3M grid given a value of \code{d}
}
\examples{

# Calculate length given a d of 6 kms
calculate_length(d = 6)

}
