% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_n.R
\name{calculate_n}
\alias{calculate_n}
\title{Function to calculate the number of hexagonal grids}
\usage{
calculate_n(x, d = NULL, area = NULL, country)
}
\arguments{
\item{x}{A class SpatialPolygons or SpatialPolygonsDataFrame object defining
the area to be sampled}

\item{d}{A numeric value for length (in kilometres) of the maximum distance
of a village/community from a sampling point.}

\item{area}{A numeric value for area (in square kilometres) of a hexagon in a
hexagonal grid defining the sampling spatial resolution}

\item{country}{Name of country where sampling area is located. This is used
to determine the appropriate UTM projection to transform \code{x}}
}
\value{
A numeric value for number of hexagons to divide sampling area based
    on a specified \code{d} or a specified hexagon \code{area} size.
}
\description{
Function to calculate the number of hexagonal grids
}
\examples{
# Calculate number of hexagons to create a sampling grid with a d = 10 kms
# on Sennar State in Sudan
sennar <- subset(sudan01, STATE = "Sennar")
calculate_n(x = sennar, d = 10, country = "Sudan")

}
