% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{countryCentroid}
\alias{countryCentroid}
\title{countryCentroid}
\format{A data.frame with 252 rows and 3 columns
\describe{
\item{\code{country}}{Name of country}
\item{\code{lon}}{Longitude coordinate of country centroid}
\item{\code{lat}}{Latitude coordinate of country centroid}
}}
\usage{
countryCentroid
}
\description{
A dataframe of country centroids
}
\keyword{datasets}
