% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_buffer.R
\name{create_buffer}
\alias{create_buffer}
\title{create_buffer}
\usage{
create_buffer(x, buffer, country)
}
\arguments{
\item{x}{A class SpatialPolygons or SpatialPolygonsDataFrame object defining
the area to be buffered}

\item{buffer}{A numeric value for distance (in kilometres) to expand the
borders of the given spatial object \code{x}. Negative values allowed.}

\item{country}{Name of country where sampling area is located. This is used
to determine the appropriate UTM projection to transform \code{x}}
}
\value{
A SpatialPolygons with the same CRS as \code{x}
}
\description{
Function to create a buffer around a SpatialPolygons or SpatialPolygonsDataFrame
object. This is a wrapper for \code{gBuffer()} function from package \code{rgeos}
adding transformation of planar coordinates to appropriate projected coordinates
to enable creation of buffer.
}
\examples{
#

}
