% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{create_sp_grid}
\alias{create_sp_grid}
\title{create_sp_grid}
\usage{
create_sp_grid(x, n, n.buffer = 0, type, fixed = FALSE)
}
\arguments{
\item{x}{Spatial object to grid}

\item{n}{Approximate number of sampling points / clusters needed. Approximate
because \code{spsample()} does not always result in \code{n} grids. If
\code{n} is the minimum number of sampling points required, specify
\code{n.buffer} to inflate \code{n} with to ensure required minimum. If
fixed \code{n} amount of sampling points needed, specify \code{fixed} to
TRUE and specify \code{n.buffer} to inflate \code{n}.}

\item{n.buffer}{Inflation factor for \code{n}. Default value is 0 (no inflation)
which produces approximate number of grids (i.e., more or less \code{n}).
If minimum number of sampling points required, specify \code{n.buffer} to
increase \code{n} with to ensure that number of grids will be at least
\code{n}. If fixed number of sampling points required (i.e., exactly \code{n}),
specify \code{n.buffer} to increase \code{n} with. Function will go
through repeat cycles of \code{spsample()} until \code{n} sampling
points are selected.}

\item{type}{A character value of type of spatial sampling to perform. Can be
one of seven options: 1) \code{"random"} - completely spatial random;
2) \code{"regular"} - systematically aligned sampling (CSAS);
3) \code{"stratified"} - stratified random (ESAS); 4) \code{"non-aligned"} -
non-aligned systematic sampling; 5) \code{"hexagonal"} - sampling on a
hexagonal lattice (S3M); 6) \code{"clustered"} - clustered sampling; and,
7) \code{"Fibonacci"} - Fibonacci sampling on the sphere.}

\item{fixed}{Logical. If TRUE, \code{n.buffer} must be specified and function
will go through repeat cycles of \code{spsample()} until \code{n}
sampling points are selected. Default is FALSE.}
}
\value{
An object of SpatialPoints-class. The number of points is only
    guaranteed to equal \code{n} when sampling is done in a square box,
    i.e. (sample.Spatial). Otherwise, the obtained number of points will have
    expected value \code{n}.
}
\description{
Function to create sampling grids as per CSAS or S3M specifications. This is
a wrapper function for \code{spsample()} function in \code{sp} package.
}
\examples{
#

}
