% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_utm_hemisphere.R
\name{find_utm_hemisphere}
\alias{find_utm_hemisphere}
\title{Find UTM hemisphere based on a longitude and latitude GPS location.}
\usage{
find_utm_hemisphere(lat)
}
\arguments{
\item{lat}{Latitude coordinate}
}
\value{
Either \code{north} or \code{south}
}
\description{
Find UTM hemisphere based on a longitude and latitude GPS location.
}
\examples{
#

}
