% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_utm_zone.R
\name{find_utm_zone}
\alias{find_utm_zone}
\title{find_utm_zone}
\usage{
find_utm_zone(lon, lat)
}
\arguments{
\item{lon}{Longitude coordinate}

\item{lat}{Latitude coordinate}
}
\value{
Numeric value for UTM zone
}
\description{
Find the UTM zone based on a longitude and latitude GPS location.
}
\examples{
#

}
