% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gedaref_villages}
\alias{gedaref_villages}
\title{A dataset of village names of Gedaref State, Sudan with data on coordinate locations}
\format{
A data frame with 5 columns and 625 rows:
\describe{
\item{\code{id}}{Unique identifier}
\item{\code{x}}{Longitude (in decimal format)}
\item{\code{y}}{Latitude (in decimal format)}
\item{\code{village}}{Village name}
\item{\code{locality}}{Locality name}
}
}
\source{
Sudan Bureau of Statistics and UNICEF Sudan
}
\usage{
gedaref_villages
}
\description{
A dataset of village names of Gedaref State, Sudan with data on coordinate locations
}
\keyword{datasets}
