% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest_point.R
\name{get_nearest_point}
\alias{get_nearest_point}
\title{get_nearest_point}
\usage{
get_nearest_point(data, x, y, query)
}
\arguments{
\item{data}{An input data frame of village/community locations with at least
information on longitude and latitude coordinates. This argument is
optional. If provided, must specify \code{x} and \code{y} parameters to
identify columns in data for longitude and latitude information}

\item{x}{A numeric vector providing longitude of village/community locations
(default). If \code{data} is specified, \code{x} must be a character value
specifying the variable name in \code{data} holding the longitude
information of the village/community locations}

\item{y}{A numeric vector providing latitude of village/community locations
(default). If \code{data} is specified, \code{y} must be a character value
specifying the variable name in \code{data} holding the latitude
information of the village/community locations}

\item{query}{A query matrix of sampling locations with at least information
on longitude and latitude coordinates. Currently, the default expected
\code{query} is that of an object of class \code{SpatialPoints} produced
by \code{spsample()} in the \code{gstat} package with longitude and
latitude data found at second and third columns}
}
\value{
A data frame of selected nearest sampling village/community locations
}
\description{
Function to select nearest community to a given sampling location
}
\examples{
#

}
\author{
Farah Mohamad Ibrahim <abdu.ff@gmail.com>
}
