% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tri.R
\name{get_tri}
\alias{get_tri}
\title{A wrapper function to \code{gDelaunayTriangulation()} to perform Delaunay
triangulation with additional options to specify thresholds for suppressing
long sided triangles}
\usage{
get_tri(input, x, y,
  crs = "+proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0",
  qTSL = 0.975)
}
\arguments{
\item{input}{A data frame of locations with at least information on longitude
and latitude coordinates of the locations.}

\item{x}{A character value specifying the variable name of the longitude
information in \code{input}.}

\item{y}{A character value specifying the variable name of the latitude
information in \code{input}.}

\item{crs}{A character value specifying the coordinate reference system to
be used for Delaunay triangulation. Default is longlat projection with
datum WGS84 (\code{"+proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0"})}

\item{qTSL}{Maximum side length to be allowed to draw sides of triangles.
Value can be between 0 to 1. Default value is 0.975.}
}
\value{
A SpatialPolygons object of Delaunay triangulation of
    \code{input} data
}
\description{
A wrapper function to \code{gDelaunayTriangulation()} to perform Delaunay
triangulation with additional options to specify thresholds for suppressing
long sided triangles
}
\examples{
#

}
