% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_utm.R
\name{get_utm}
\alias{get_utm}
\title{get_utm}
\usage{
get_utm(lon, lat, units = "m")
}
\arguments{
\item{lon}{Longitude coordinate}

\item{lat}{Latitude coordinate}

\item{units}{UTM measurements units. Default is \code{m} for metres.}
}
\value{
A character string specifying the appropriate UTM projection
}
\description{
Get the appropriate UTM projection string given a longitude and latitude
coordinate reference system values.
}
\examples{
#

}
