% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{map_projections}
\alias{map_projections}
\title{A dataset of UTM projections}
\format{
A data.frame with 12 rows and 3 columns:
\describe{
\item{\code{country}}{Name of country}
\item{\code{utm}}{UTM name}
\item{\code{proj}}{A character string specifying UTM projection}
}
}
\usage{
map_projections
}
\description{
A dataset of UTM projections
}
\keyword{datasets}
