% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_triangles.R
\name{plot_tri}
\alias{plot_tri}
\title{plot_tri}
\usage{
plot_tri(x, border, qTSL)
}
\arguments{
\item{x}{A SpatialPolygonsDataFrame object of \code{Delaunay triangulation}
output of function \code{gDelaunayTriangulation()}}

\item{border}{Colour of border to use when plotting the triangles}

\item{qTSL}{Maximum side length to be allowed to draw sides of triangles.
Value can be between 0 to 1. Default value is 0.975.}
}
\value{
Delaunay triangulation of selected sampling villages
}
\description{
Function to suppress drawing long-sided triangles
}
\examples{
#

}
\author{
Farah Mohamad Ibrahim <abdu.ff@gmail.com>
}
