% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3m.R
\docType{package}
\name{s3m}
\alias{s3m}
\alias{s3m-package}
\title{Simple Spatial Sampling Method (S3M)}
\description{
The Simple Spatial Survey Method (S3M) was developed from the Centric
Systematic Area Sampling (CSAS) approach used in biodiversity species
counting and adapted for use in health and nutrition surveys by Brixton Health
and Valid International. S3M is used in surveys designed to be spatially
representative i.e., the sample distributed evenly across the survey area,
using a spatial sample design that selected communities located closest to
the centroids of a hexagonal grid laid over the survey area. The resulting
sample is a triangular irregular network.
}
