% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sennar_villages}
\alias{sennar_villages}
\title{A dataset of village names of Sennar State, Sudan with data on coordinate locations}
\format{A data frame with 5 columns and 741 rows:
\describe{
\item{\code{id}}{Unique identifier}
\item{\code{x}}{Longitude (in decimal format)}
\item{\code{y}}{Latitude (in decimal format)}
\item{\code{village}}{Village name}
\item{\code{locality}}{Locality name}
}}
\source{
Sudan Bureau of Statistics and UNICEF Sudan

Sudan Bureau of Statistics and UNICEF Sudan
}
\usage{
sennar_villages

sennar_villages
}
\description{
A dataset of village names of Sennar State, Sudan with data on coordinate locations

A dataset of village names of Sennar State, Sudan with data on coordinate locations
}
\keyword{datasets}
