% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{turkey00}
\alias{turkey00}
\title{turkey00}
\format{A data frame with 4 columns and 1 rows:  
\describe{ 
\item{\code{NAME_EN}}{Name (English)}  
\item{\code{ADM0}}{Administrative level 0 Code}  
\item{\code{Shape_Leng}}{Length of Shape}  
\item{\code{Shape_Area}}{Area of Shape}  
}}
\source{
General Command of Mapping of Turkey (\url{http://www.hgk.msb.gov.tr/u-23-turkiye-mulki-idare-sinirlari.html})
}
\usage{
turkey00
}
\description{
A map dataset of class SpatialPolygonsDataFrame of level 0 administrative units (national) of Turkey
}
\keyword{datasets}
