% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{turkey01}
\alias{turkey01}
\title{turkey01}
\format{A data frame with 6 columns and 81 rows:  
\describe{ 
\item{\code{Shape_Leng}}{Length of Shape}  
\item{\code{Shape_Area}}{Area of Shape}  
\item{\code{NAME_EN}}{Province name (English)}  
\item{\code{ADM0}}{Administrative level 0 Code}  
\item{\code{ADM1}}{Administrative level 1 Code}  
\item{\code{PCODE}}{Postal Code}  
}}
\source{
General Command of Mapping of Turkey (\url{http://www.hgk.msb.gov.tr/u-23-turkiye-mulki-idare-sinirlari.html})
}
\usage{
turkey01
}
\description{
A map dataset of class SpatialPolygonsDataFrame of level 1 administrative units (province) of Turkey
}
\keyword{datasets}
