% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{turkey02}
\alias{turkey02}
\title{turkey02}
\format{A data frame with 10 columns and 970 rows:  
\describe{ 
\item{\code{NAME_EN}}{District Name (English)}  
\item{\code{PROV_EN}}{Province Name (English)}  
\item{\code{ADM0}}{Administrative level 0 Code}  
\item{\code{ADM1}}{Administrative level 1 Code}  
\item{\code{ADM2}}{Administrative level 2 Code}  
\item{\code{PCODE}}{Postal Code}  
\item{\code{Shape_Leng}}{Length of Shape}  
\item{\code{ADDINFO}}{Additional Information}  
\item{\code{Shape_Le_1}}{Length of Shape}  
\item{\code{Shape_Area}}{Area of Shape}  
}}
\source{
General Command of Mapping of Turkey (\url{http://www.hgk.msb.gov.tr/u-23-turkiye-mulki-idare-sinirlari.html})
}
\usage{
turkey02
}
\description{
A map dataset of class SpatialPolygonsDataFrame of level 2 administrative units (district) of Turkey
}
\keyword{datasets}
