% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vary_density.R
\name{vary_density}
\alias{vary_density}
\title{vary_density}
\usage{
vary_density(sample, x, y, label = seq_along(x), basemap = NULL,
  crs = "+proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0")
}
\arguments{
\item{sample}{A data frame containing selected villages for sampling. This is the
data frame produced by \code{get_nearest_point()} function.}

\item{x}{A character value specifying the variable name in \code{sample}
holding the longitude information of the sampling locations.}

\item{y}{A character value specifying the variable name in \code{sample}
holding the latitude information of the sampling locations.}

\item{label}{A character value specifying the variable name in \code{sample}
holding the name of the sampling locations. If no name information
available in data frame, specify a character vector with length equal to
number of sampling locations in \code{sample}. Default is NULL.}

\item{basemap}{A SpatialPolygonsDataFrame of base map to show when plotting
triangulation. Default is NULL}

\item{crs}{A character vector specifying the coordinate reference system to
use for the spatial objects. Default is longlat projection for datum
WGS84}
}
\description{
Adjust triangulation by by dividing a densely populated triangle into smaller
triangles by adding 3 points at the middle of each side of the triangle
}
\examples{
#

}
