% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest_point.R
\name{get_nearest_point}
\alias{get_nearest_point}
\title{get_nearest_point}
\usage{
get_nearest_point(input, x1, y1, query, x2, y2, n = 1)
}
\arguments{
\item{input}{A matrix or data frame of input sampling locations to which nearest
village locations are to be matched. Data frame should contain at least
information on longitude and latitude coordinates.}

\item{x1}{A character value specifying the variable name in \code{input}
holding the longitude information of the sampling locations}

\item{y1}{A character value specifying the variable name in \code{input}
holding the latitude information of the village/community locations}

\item{query}{A data frame of village/community locations with at least
information on longitude and latitude coordinates from which to query for
nearest point.}

\item{x2}{A character value specifying the variable name in \code{query}
holding the longitude information of the village/community locations}

\item{y2}{A character value specifying the variable name in \code{query}
holding the latitude information of the village/community locations}

\item{n}{Number of nearest village/community locations to select}
}
\value{
A data frame of selected nearest sampling village/community locations
}
\description{
Function to select nearest community to a given sampling location
}
\examples{
#

}
\author{
Farah Mohamad Ibrahim <abdu.ff@gmail.com>
}
